rename project_comms to message_handlers
