+++ example/firmware/config/message_handlers.h
