+++ example/firmware/spi_example.c
@@ -6,7 +6,7 @@
 #include "control.h"
 #include "frequency.h"
 #include "priority.h"
-#include "project_comms.h"
+#include "message_handlers.h"
 #include <timer_pcbDetails.h>
